Attempt to make header margins more like original design for Contact Information section.
Since the hcard microformat is no longer being used on the Luther sites in the Contact Information section, the original styling was a bit broken.  This change removes bottom margins on H2 headers and adds it back to the first H2 that is a direct child of .contact-info.

Note: .contact-info > h2:first-of-type may not work in IE7 or less.